refactor: 💡 OAuth 관련 패키지를 user 하위에서 web의 하위 디렉토리로 이동
- 반영한 코드 리뷰:
https://github.com/Sc0up/scoup-backend/pull/22#discussion_r686516326

✅ Closes: #12
