+++ src/main/java/com/postsquad/scoup/web/auth/OAuth.java
@@ -1,0 +1,11 @@
+package com.postsquad.scoup.web.auth;
+
+import com.postsquad.scoup.web.auth.response.AccessTokenResponse;
+import com.postsquad.scoup.web.auth.response.OAuthUserResponse;
+
+public interface OAuth {
+
+    AccessTokenResponse getToken(String code);
+
+    OAuthUserResponse getUserInfo(String token);
+}
